Extract deployment secret name generation into the names package
