+++ pkg/bosh/manifest/kube_converter.go
@@ -4,7 +4,6 @@ import (
 	"crypto/sha1"
 	"fmt"
 	"path/filepath"
-	"regexp"
 	"strings"
 
 	"github.com/pkg/errors"
@@ -18,6 +17,7 @@ import (
 	ejv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedjob/v1alpha1"
 	esv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedsecret/v1alpha1"
 	essv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/extendedstatefulset/v1alpha1"
+	"code.cloudfoundry.org/cf-operator/pkg/kube/util/names"
 )
 
 const (
@@ -104,7 +104,7 @@ func (m *Manifest) variableInterpolationJob(namespace string) (*ejv1.ExtendedJob
 	args := []string{"-c", `cf-operator variable-interpolation`}
 
 	// This is the source manifest, that still has the '((vars))'
-	manifestSecretName := m.CalculateSecretName(DeploymentSecretTypeManifestWithOps, "")
+	manifestSecretName := names.CalculateSecretName(names.DeploymentSecretTypeManifestWithOps, m.Name, "")
 
 	// Prepare Volumes and Volume mounts
 
@@ -136,7 +136,7 @@ func (m *Manifest) variableInterpolationJob(namespace string) (*ejv1.ExtendedJob
 	}
 	for _, variable := range variables {
 		varName := variable
-		varSecretName := m.CalculateSecretName(DeploymentSecretTypeGeneratedVariable, varName)
+		varSecretName := names.CalculateSecretName(names.DeploymentSecretTypeGeneratedVariable, m.Name, varName)
 
 		// The volume definition
 		vol := v1.Volume{
@@ -184,8 +184,9 @@ func (m *Manifest) variableInterpolationJob(namespace string) (*ejv1.ExtendedJob
 		return nil, errors.Wrap(err, "could not calculate manifest SHA1")
 	}
 
-	outputSecretPrefix, _ := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeManifestAndVars,
+	outputSecretPrefix, _ := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeManifestAndVars,
+		m.Name,
 		VarInterpolationContainerName,
 		false,
 	)
@@ -263,15 +264,17 @@ func (m *Manifest) SHA1() (string, error) {
 // dataGatheringJob generates the Data Gathering Job for a manifest
 func (m *Manifest) dataGatheringJob(namespace string) (*ejv1.ExtendedJob, error) {
 
-	_, interpolatedManifestSecretName := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeManifestAndVars,
+	_, interpolatedManifestSecretName := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeManifestAndVars,
+		m.Name,
 		VarInterpolationContainerName,
 		true,
 	)
 
 	eJobName := fmt.Sprintf("data-gathering-%s", m.Name)
-	outputSecretNamePrefix, _ := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeInstanceGroupResolvedProperties,
+	outputSecretNamePrefix, _ := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeInstanceGroupResolvedProperties,
+		m.Name,
 		"",
 		false,
 	)
@@ -418,8 +421,9 @@ func (m *Manifest) jobsToInitContainers(igName string, jobs []Job, namespace str
 
 	}
 
-	_, resolvedPropertiesSecretName := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeInstanceGroupResolvedProperties,
+	_, resolvedPropertiesSecretName := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeInstanceGroupResolvedProperties,
+		m.Name,
 		igName,
 		true,
 	)
@@ -510,13 +514,15 @@ func (m *Manifest) serviceToExtendedSts(ig *InstanceGroup, namespace string) (es
 		return essv1.ExtendedStatefulSet{}, err
 	}
 
-	_, interpolatedManifestSecretName := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeManifestAndVars,
+	_, interpolatedManifestSecretName := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeManifestAndVars,
+		m.Name,
 		VarInterpolationContainerName,
 		true,
 	)
-	_, resolvedPropertiesSecretName := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeInstanceGroupResolvedProperties,
+	_, resolvedPropertiesSecretName := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeInstanceGroupResolvedProperties,
+		m.Name,
 		ig.Name,
 		true,
 	)
@@ -619,13 +625,15 @@ func (m *Manifest) errandToExtendedJob(ig *InstanceGroup, namespace string) (ejv
 		return ejv1.ExtendedJob{}, err
 	}
 
-	_, interpolatedManifestSecretName := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeManifestAndVars,
+	_, interpolatedManifestSecretName := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeManifestAndVars,
+		m.Name,
 		VarInterpolationContainerName,
 		true,
 	)
-	_, resolvedPropertiesSecretName := m.CalculateEJobOutputSecretPrefixAndName(
-		DeploymentSecretTypeInstanceGroupResolvedProperties,
+	_, resolvedPropertiesSecretName := names.CalculateEJobOutputSecretPrefixAndName(
+		names.DeploymentSecretTypeInstanceGroupResolvedProperties,
+		m.Name,
 		ig.Name,
 		true,
 	)
@@ -704,7 +712,7 @@ func (m *Manifest) convertVariables(namespace string) []esv1.ExtendedSecret {
 	secrets := []esv1.ExtendedSecret{}
 
 	for _, v := range m.Variables {
-		secretName := m.CalculateSecretName(DeploymentSecretTypeGeneratedVariable, v.Name)
+		secretName := names.CalculateSecretName(names.DeploymentSecretTypeGeneratedVariable, m.Name, v.Name)
 		s := esv1.ExtendedSecret{
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      secretName,
@@ -726,11 +734,11 @@ func (m *Manifest) convertVariables(namespace string) []esv1.ExtendedSecret {
 			}
 			if v.Options.CA != "" {
 				certRequest.CARef = esv1.SecretReference{
-					Name: m.CalculateSecretName(DeploymentSecretTypeGeneratedVariable, v.Options.CA),
+					Name: names.CalculateSecretName(names.DeploymentSecretTypeGeneratedVariable, m.Name, v.Options.CA),
 					Key:  "certificate",
 				}
 				certRequest.CAKeyRef = esv1.SecretReference{
-					Name: m.CalculateSecretName(DeploymentSecretTypeGeneratedVariable, v.Options.CA),
+					Name: names.CalculateSecretName(names.DeploymentSecretTypeGeneratedVariable, m.Name, v.Options.CA),
 					Key:  "private_key",
 				}
 			}
@@ -888,47 +896,4 @@ func (m *Manifest) JobSpecCopierContainer(releaseName string, releaseImage strin
 	}
 
 	return initContainers
-}
-
-// AllVariableNames returns the names of all variables
-func (m *Manifest) AllVariableNames() ([]string, error) {
-	varMap := make(map[string]struct{})
-
-	// Add all "explicit" variables from the "variables" block
-	for _, v := range m.Variables {
-		varMap[v.Name] = struct{}{}
-	}
-
-	// Find all properties that could contain "implicit" variables like "app_domain: ((system_domain))"
-	properties := []interface{}{}
-	for _, ig := range m.InstanceGroups {
-		for _, property := range ig.Properties {
-			properties = append(properties, property)
-		}
-
-		for _, job := range ig.Jobs {
-			for _, property := range job.Properties.Properties {
-				properties = append(properties, property)
-			}
-		}
-	}
-
-	// Extract "implicit" variables
-	r := regexp.MustCompile(`\(\((.*)\)\)`)
-	for _, property := range properties {
-		raw, err := yaml.Marshal(property)
-		if err != nil {
-			return []string{}, errors.Wrapf(err, "marshalling property '%s' failed", property)
-		}
-		for _, match := range r.FindAllStringSubmatch(string(raw), -1) {
-			varMap[match[1]] = struct{}{}
-		}
-	}
-
-	names := []string{}
-	for k := range varMap {
-		names = append(names, k)
-	}
-
-	return names, nil
 }
