+++ pkg/kube/util/names/names.go
@@ -1,6 +1,7 @@
 package names
 
 import (
+	"crypto/md5"
 	"crypto/rand"
 	"encoding/hex"
 	"fmt"
@@ -11,6 +12,72 @@ import (
 
 	"github.com/pkg/errors"
 )
+
+// DeploymentSecretType lists all the types of secrets used in
+// the lifecycle of a BOSHDeployment
+type DeploymentSecretType int
+
+const (
+	// DeploymentSecretTypeManifestWithOps is a manifest that has ops files applied
+	DeploymentSecretTypeManifestWithOps DeploymentSecretType = iota
+	// DeploymentSecretTypeManifestAndVars is a manifest whose variables have been interpolated
+	DeploymentSecretTypeManifestAndVars
+	// DeploymentSecretTypeGeneratedVariable is a BOSH variable generated using an ExtendedSecret
+	DeploymentSecretTypeGeneratedVariable
+	// DeploymentSecretTypeInstanceGroupResolvedProperties is a YAML file containing all properties needed to render an Instance Group
+	DeploymentSecretTypeInstanceGroupResolvedProperties
+	// DeploymentSecretTypeImplicitVariable is a BOSH variable provided by the user as a Secret
+	DeploymentSecretTypeImplicitVariable
+)
+
+func (s DeploymentSecretType) String() string {
+	return [...]string{
+		"with-ops",
+		"with-vars",
+		"var",
+		"ig-resolved",
+		"implicit-var"}[s]
+}
+
+// CalculateSecretName generates a Secret name for a given name and a deployment
+func CalculateSecretName(secretType DeploymentSecretType, deploymentName, name string) string {
+	if name == "" {
+		name = secretType.String()
+	} else {
+		name = fmt.Sprintf("%s-%s", secretType, name)
+	}
+
+	nameRegex := regexp.MustCompile("[^-][a-z0-9-]*.[a-z0-9-]*[^-]")
+	partRegex := regexp.MustCompile("[a-z0-9-]*")
+
+	deploymentName = partRegex.FindString(strings.Replace(deploymentName, "_", "-", -1))
+	variableName := partRegex.FindString(strings.Replace(name, "_", "-", -1))
+	secretName := nameRegex.FindString(deploymentName + "." + variableName)
+
+	if len(secretName) > 63 {
+		// secret names are limited to 63 characters so we recalculate the name as
+		// <name trimmed to 31 characters><md5 hash of name>
+		sumHex := md5.Sum([]byte(secretName))
+		sum := hex.EncodeToString(sumHex[:])
+		secretName = secretName[:63-32] + sum
+	}
+
+	return secretName
+}
+
+// CalculateEJobOutputSecretPrefixAndName generates a Secret prefix for the output
+// of an Extended Job given a name, and calculates the final Secret name,
+// given a container name
+func CalculateEJobOutputSecretPrefixAndName(secretType DeploymentSecretType, deploymentName string, containerName string, versioned bool) (string, string) {
+	prefix := CalculateSecretName(secretType, deploymentName, "")
+	finalName := fmt.Sprintf("%s.%s", prefix, containerName)
+
+	if versioned {
+		finalName = fmt.Sprintf("%s-v0", finalName)
+	}
+
+	return prefix + ".", finalName
+}
 
 // GetStatefulSetName gets statefulset name from podName
 func GetStatefulSetName(name string) string {
